Added interactions endpoint
