--- src/main/java/com/hype/eventservice/api/action/domain/ActionRepository.java
@@ -1,8 +1,0 @@
-package com.hype.eventservice.api.action.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.math.BigInteger;
-
-public interface ActionRepository extends JpaRepository<Action, BigInteger> {
-}
