+++ src/main/java/com/hype/eventservice/api/backoffice/common/domain/ActionRepository.java
@@ -1,0 +1,8 @@
+package com.hype.eventservice.api.backoffice.common.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.math.BigInteger;
+
+public interface ActionRepository extends JpaRepository<Action, BigInteger> {
+}
