+++ src/main/java/com/hype/eventservice/api/backoffice/common/domain/UserRepository.java
@@ -1,0 +1,6 @@
+package com.hype.eventservice.api.backoffice.common.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, String> {
+}
