--- src/main/java/com/hype/eventservice/api/event/domain/EventRepository.java
@@ -1,8 +1,0 @@
-package com.hype.eventservice.api.event.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.math.BigInteger;
-
-public interface EventRepository extends JpaRepository<Event, BigInteger> {
-}
