--- src/main/java/com/hype/eventservice/api/event/dto/RestResponseDTO.java
@@ -1,14 +1,0 @@
-package com.hype.eventservice.api.event.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class RestResponseDTO {
-
-    private String status;
-    private String message;
-}
