+++ src/main/resources/application.properties
@@ -5,9 +5,5 @@ spring.datasource.password=${DATASOURCE_PASSWORD}
 spring.datasource.driver-class=com.mysql.jdbc.Driver
 spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
 spring.jpa.show-sql = true
-#spring.jpa.generate-ddl=true
-#spring.jpa.hibernate.ddl-auto=update
-
-security.oauth2.client.scope=password
-security.oauth2.client.client-id=${CLIENT_ID}
-security.oauth2.client.client-secret=${CLIENT_SECRET}
+spring.jpa.generate-ddl=true
+spring.jpa.hibernate.ddl-auto=update
